add a clock abstraction to make tests faster
There are various timeouts involved in dealing STUN requests that, while
we could wait the full amount of wall clock time, that's a lot of
sitting around doing nothing in a test.

What we can do instead for tests is implement our own clock that can time
travel as required.