conncheck: fix a race when a stun request creates the StunAgent (e.g. TCP)
The callback mechanism we have implemented with channels may not
actually have successfully set the agent on the StunRequest when we go
to ask for the `created_agent()` from it.  Fix by keeping the reference
to the StunAgent in the conncheck earlier when the Stun Request asks the
conncheck to configure the StunAgent.