conncheck: fix handling for duplicate triggered checks
A check being triggered twice may result in the second trigger producing
a large wait that may never be completed.  That is not ideal and we
instead assert that this must not happen with a panic().

Then fixing the panic involves not triggering a check if it is already
has an associated STUN request.